support building against Octave.app, add octapp 9.1b4 and 9.2b1 support
This code organization is poor, but it at least gets a basic build to run. I don't know if it's really "correct", because it's just pointing a couple paths at Octave.app, and not pulling in its entire bundled build toolchain into the environment by setting $PATH and/or other env vars. Which we should probably do for a "correct" build environment, but I don't know if it actually matters in this particular test case.